Move connection save/duplicate model logic into ConnectionManager
ConnectionManager gains generate_duplicate_nickname, duplicate_connection,
apply_connection_update, and create_connection; MainWindow keeps only the
UI side (list rebuild, row selection, group/meta migration, reconnect
prompt, dialogs). Drops the dead change-detection dict and the duplicated
reload block in on_connection_saved, and replaces the raw Gtk.MessageDialog
error with the existing _error_dialog helper.
