+++ sshpilot/connection_manager.py
@@ -3,6 +3,7 @@ Connection Manager for sshPilot
 Handles SSH connections, configuration, and secure password storage
 """
 
+import copy
 import os
 import stat
 import shutil
@@ -14,6 +15,7 @@ import getpass
 import subprocess
 import shlex
 import re
+from gettext import gettext as _
 from pathlib import Path
 from typing import Dict, List, Optional, Any, Tuple, Union, Set
 
@@ -2948,3 +2950,355 @@ class ConnectionManager(GObject.Object):
             if connection.nickname == nickname:
                 return connection
         return None
+
+    def generate_duplicate_nickname(self, base_nickname: str) -> str:
+        """Generate a unique nickname for a duplicated connection."""
+        try:
+            existing_names = {
+                str(getattr(conn, 'nickname', '')).strip()
+                for conn in self.get_connections()
+                if getattr(conn, 'nickname', None)
+            }
+        except Exception:
+            existing_names = set()
+        existing_lower = {name.lower() for name in existing_names if name}
+
+        base = (base_nickname or '').strip()
+        if not base:
+            base = _('Connection')
+
+        copy_label = _('Copy')
+        # The nickname is used verbatim as the ssh Host alias, and the app's own
+        # validator rejects whitespace (and parens make an invalid host token —
+        # see #953). So the suffix must be whitespace/paren-free: use a hyphen
+        # separator and a whitespace-free copy token, e.g. "Name-Copy",
+        # "Name-Copy-2".
+        copy_token = re.sub(r"\s+", "-", copy_label.strip()) or "Copy"
+        # Strip an existing copy suffix in either the legacy " (Copy[ N])" form
+        # or the new "-Copy[-N]" form so re-duplicating doesn't stack suffixes.
+        pattern = re.compile(
+            r"(?:\s*\(\s*" + re.escape(copy_label) + r"(?:\s+\d+)?\s*\)"
+            r"|[-_]+" + re.escape(copy_token) + r"(?:[-_]+\d+)?)\s*$",
+            re.IGNORECASE,
+        )
+        base_clean = pattern.sub('', base).strip() or base
+
+        def is_unique(name: str) -> bool:
+            return name.lower() not in existing_lower
+
+        candidate = f"{base_clean}-{copy_token}"
+        if is_unique(candidate):
+            return candidate
+
+        index = 2
+        while True:
+            candidate = f"{base_clean}-{copy_token}-{index}"
+            if is_unique(candidate):
+                return candidate
+            index += 1
+
+    def duplicate_connection(self, connection: Connection, group_manager) -> Connection:
+        """Duplicate *connection*, persist it, and mirror its group placement.
+
+        Returns the duplicated :class:`Connection`; raises on failure. UI
+        concerns (list rebuild, row selection, error dialogs) are the caller's.
+        """
+        try:
+            base_data = getattr(connection, 'data', None)
+            new_data = copy.deepcopy(base_data) if isinstance(base_data, dict) else {}
+        except Exception:
+            new_data = {}
+        if not isinstance(new_data, dict):
+            new_data = {}
+
+        for key in list(new_data.keys()):
+            if key.startswith('__') or key in {'aliases', 'password_changed'}:
+                new_data.pop(key, None)
+
+        new_nickname = self.generate_duplicate_nickname(getattr(connection, 'nickname', ''))
+        new_data['nickname'] = new_nickname
+
+        # Plugin protocols: the data dict is authoritative (no ssh_config
+        # attribute fixups apply); persist through the non-SSH store.
+        if getattr(connection, 'protocol', 'ssh') != 'ssh':
+            new_connection = Connection(new_data)
+            if not self.update_connection(new_connection, dict(new_data)):
+                raise RuntimeError(_('Failed to save duplicated connection.'))
+            original_groups = group_manager.get_connection_groups(connection.nickname)
+            original_group_id = original_groups[0] if original_groups else None
+            if original_group_id and original_group_id in getattr(group_manager, 'groups', {}):
+                try:
+                    group_manager.move_connection(new_nickname, original_group_id)
+                except Exception:
+                    pass
+            return new_connection
+
+        host_value = (
+            getattr(connection, 'hostname', '')
+            or getattr(connection, 'host', '')
+            or new_data.get('hostname', '')
+            or new_data.get('host', '')
+        )
+        host_value = str(host_value).strip()
+        if not host_value:
+            host_value = new_nickname
+        new_data['hostname'] = host_value
+        new_data.pop('host', None)
+
+        new_data['username'] = str(getattr(connection, 'username', new_data.get('username', '')) or '')
+
+        try:
+            new_data['port'] = int(getattr(connection, 'port', new_data.get('port', 22)) or 22)
+        except Exception:
+            new_data['port'] = 22
+
+        try:
+            new_data['auth_method'] = int(getattr(connection, 'auth_method', new_data.get('auth_method', 0)) or 0)
+        except Exception:
+            new_data['auth_method'] = 0
+
+        keyfile_value = getattr(connection, 'keyfile', new_data.get('keyfile', '')) or ''
+        if isinstance(keyfile_value, str) and keyfile_value.strip().lower().startswith('select key file'):
+            keyfile_value = ''
+        new_data['keyfile'] = keyfile_value
+
+        certificate_value = getattr(connection, 'certificate', new_data.get('certificate', '')) or ''
+        if isinstance(certificate_value, str) and certificate_value.strip().lower().startswith('select certificate'):
+            certificate_value = ''
+        new_data['certificate'] = certificate_value
+
+        new_data['key_passphrase'] = getattr(connection, 'key_passphrase', new_data.get('key_passphrase', '')) or ''
+
+        try:
+            new_data['key_select_mode'] = int(getattr(connection, 'key_select_mode', new_data.get('key_select_mode', 0)) or 0)
+        except Exception:
+            new_data['key_select_mode'] = 0
+
+        new_data['password'] = getattr(connection, 'password', new_data.get('password', '')) or ''
+        new_data['x11_forwarding'] = bool(getattr(connection, 'x11_forwarding', new_data.get('x11_forwarding', False)))
+        new_data['pubkey_auth_no'] = bool(getattr(connection, 'pubkey_auth_no', new_data.get('pubkey_auth_no', False)))
+        new_data['forward_agent'] = bool(getattr(connection, 'forward_agent', new_data.get('forward_agent', False)))
+
+        proxy_jump_value = getattr(connection, 'proxy_jump', new_data.get('proxy_jump', []))
+        if isinstance(proxy_jump_value, str):
+            proxy_jump_value = [h.strip() for h in re.split(r'[\s,]+', proxy_jump_value) if h.strip()]
+        else:
+            proxy_jump_value = list(proxy_jump_value or [])
+        new_data['proxy_jump'] = proxy_jump_value
+
+        new_data['proxy_command'] = getattr(connection, 'proxy_command', new_data.get('proxy_command', '')) or ''
+        new_data['pre_command'] = getattr(connection, 'pre_command', new_data.get('pre_command', '')) or ''
+        new_data['local_command'] = getattr(connection, 'local_command', new_data.get('local_command', '')) or ''
+        new_data['remote_command'] = getattr(connection, 'remote_command', new_data.get('remote_command', '')) or ''
+        new_data['extra_ssh_config'] = getattr(connection, 'extra_ssh_config', new_data.get('extra_ssh_config', '')) or ''
+
+        forwarding_rules = getattr(connection, 'forwarding_rules', new_data.get('forwarding_rules', []))
+        try:
+            new_data['forwarding_rules'] = copy.deepcopy(list(forwarding_rules or []))
+        except Exception:
+            new_data['forwarding_rules'] = []
+
+        source_path = getattr(connection, 'source', new_data.get('source'))
+        if source_path:
+            new_data['source'] = source_path
+        else:
+            new_data.pop('source', None)
+
+        new_connection = Connection(new_data)
+        if self.isolated_mode:
+            new_connection.isolated_config = True
+            new_connection.config_root = self.ssh_config_path
+            new_connection.data['isolated_mode'] = True
+            if self.ssh_config_path:
+                new_connection.data['config_root'] = self.ssh_config_path
+        try:
+            new_connection.auth_method = int(new_data.get('auth_method', 0) or 0)
+        except Exception:
+            new_connection.auth_method = 0
+        try:
+            new_connection.key_select_mode = int(new_data.get('key_select_mode', 0) or 0)
+        except Exception:
+            new_connection.key_select_mode = 0
+        new_connection.forwarding_rules = list(new_data.get('forwarding_rules', []))
+        new_connection.proxy_jump = list(new_data.get('proxy_jump', []))
+        new_connection.forward_agent = bool(new_data.get('forward_agent', False))
+        new_connection.extra_ssh_config = new_data.get('extra_ssh_config', '')
+        new_connection.certificate = new_data.get('certificate', '')
+
+        original_groups = group_manager.get_connection_groups(connection.nickname)
+        original_group_id = original_groups[0] if original_groups else None
+
+        self.connections.append(new_connection)
+        try:
+            if not self.update_connection(new_connection, new_data):
+                raise RuntimeError(_('Failed to save duplicated connection.'))
+        except Exception:
+            try:
+                self.connections.remove(new_connection)
+            except ValueError:
+                pass
+            raise
+
+        self.load_ssh_config()
+
+        if original_group_id and original_group_id in getattr(group_manager, 'groups', {}):
+            group_manager.move_connection(new_nickname, original_group_id)
+            try:
+                group_manager.reorder_connection_in_group(new_nickname, connection.nickname, 'below')
+            except Exception:
+                pass
+            # Mirror any additional group memberships of the original
+            for extra_group_id in original_groups[1:]:
+                if extra_group_id in getattr(group_manager, 'groups', {}):
+                    group_manager.copy_connection_to_group(new_nickname, extra_group_id)
+        else:
+            group_manager.move_connection(new_nickname, None)
+            try:
+                root_connections = group_manager.root_connections
+                if new_nickname in root_connections and connection.nickname in root_connections:
+                    root_connections.remove(new_nickname)
+                    insert_at = root_connections.index(connection.nickname) + 1
+                    root_connections.insert(insert_at, new_nickname)
+                    group_manager._save_groups()
+            except Exception:
+                pass
+
+        return self.find_connection_by_nickname(new_nickname) or new_connection
+
+    def apply_connection_update(self, old_connection: Connection, connection_data: Dict[str, Any]) -> bool:
+        """Persist an edited connection and sync the live instance's attributes.
+
+        Returns False when the config write fails. Group/metadata migration and
+        UI refresh are the caller's concern.
+        """
+        # Ensure auth_method always present and normalized
+        try:
+            connection_data['auth_method'] = int(connection_data.get('auth_method', getattr(old_connection, 'auth_method', 0)) or 0)
+        except Exception:
+            connection_data['auth_method'] = 0
+
+        if not self.update_connection(old_connection, connection_data):
+            return False
+
+        # Update connection attributes in memory (ensure forwarding rules kept)
+        old_connection.nickname = connection_data['nickname']
+        old_connection.hostname = connection_data['hostname']
+        old_connection.host = old_connection.hostname
+        old_connection.username = connection_data['username']
+        old_connection.port = connection_data['port']
+        old_connection.keyfile = connection_data['keyfile']
+        old_connection.certificate = connection_data.get('certificate', '')
+        old_connection.password = connection_data['password']
+        old_connection.key_passphrase = connection_data.get('key_passphrase', getattr(old_connection, 'key_passphrase', '')) or ''
+        old_connection.auth_method = connection_data['auth_method']
+        # Persist key selection mode in-memory so the dialog reflects it without restart
+        try:
+            old_connection.key_select_mode = int(connection_data.get('key_select_mode', getattr(old_connection, 'key_select_mode', 0)) or 0)
+        except Exception:
+            pass
+        old_connection.x11_forwarding = connection_data['x11_forwarding']
+        old_connection.forwarding_rules = list(connection_data.get('forwarding_rules', []))
+        # Refresh proxy settings in-memory so new connections immediately pick
+        # up the updated directives without a full application restart.
+        try:
+            proxy_jump_value = connection_data.get('proxy_jump', [])
+            if isinstance(proxy_jump_value, str):
+                proxy_jump_value = [
+                    h.strip() for h in re.split(r'[\s,]+', proxy_jump_value) if h.strip()
+                ]
+            else:
+                proxy_jump_value = [
+                    str(h).strip() for h in (proxy_jump_value or []) if str(h).strip()
+                ]
+            old_connection.proxy_jump = proxy_jump_value
+        except Exception:
+            proxy_jump_value = []
+            old_connection.proxy_jump = []
+
+        proxy_command_value = connection_data.get('proxy_command', '') or ''
+        forward_agent_value = bool(connection_data.get('forward_agent', False))
+
+        old_connection.proxy_command = proxy_command_value
+        old_connection.forward_agent = forward_agent_value
+
+        # Keep the backing data dict synchronized so any downstream consumers
+        # that still read from connection.data see the new directives without
+        # waiting for another reload cycle.
+        try:
+            if hasattr(old_connection, 'data') and isinstance(old_connection.data, dict):
+                old_connection.data['proxy_jump'] = list(proxy_jump_value)
+                old_connection.data['proxy_command'] = proxy_command_value
+                old_connection.data['forward_agent'] = forward_agent_value
+        except Exception:
+            pass
+
+        # Invalidate any prepared SSH command so future connection attempts
+        # rebuild the argument list using the refreshed proxy settings.
+        try:
+            if hasattr(old_connection, 'ssh_cmd'):
+                old_connection.ssh_cmd = []
+        except Exception:
+            try:
+                delattr(old_connection, 'ssh_cmd')
+            except Exception:
+                pass
+
+        try:
+            old_connection.pre_command = connection_data.get('pre_command', '')
+            old_connection.local_command = connection_data.get('local_command', '')
+            old_connection.remote_command = connection_data.get('remote_command', '')
+            old_connection.extra_ssh_config = connection_data.get('extra_ssh_config', '')
+        except Exception:
+            pass
+
+        return True
+
+    def create_connection(self, connection_data: Dict[str, Any]) -> Optional[Connection]:
+        """Create and persist a new connection from dialog data.
+
+        Returns the new :class:`Connection`, or None when the config write
+        fails. UI refresh is the caller's concern.
+        """
+        connection = Connection(connection_data)
+        if self.isolated_mode:
+            connection.isolated_config = True
+            connection.config_root = self.ssh_config_path
+            connection.data['isolated_mode'] = True
+            if self.ssh_config_path:
+                connection.data['config_root'] = self.ssh_config_path
+        # Ensure the in-memory object reflects the dialog's choices immediately
+        try:
+            connection.auth_method = int(connection_data.get('auth_method', 0))
+        except Exception:
+            connection.auth_method = 0
+        try:
+            connection.key_select_mode = int(connection_data.get('key_select_mode', 0) or 0)
+        except Exception:
+            connection.key_select_mode = 0
+        try:
+            connection.certificate = connection_data.get('certificate', '')
+        except Exception:
+            connection.certificate = ''
+        try:
+            connection.extra_ssh_config = connection_data.get('extra_ssh_config', '')
+        except Exception:
+            connection.extra_ssh_config = ''
+        self.connections.append(connection)
+
+        if not self.update_connection(connection, connection_data):
+            return None
+
+        # Reload from SSH config so the in-memory list matches disk
+        try:
+            self.load_ssh_config()
+        except Exception:
+            pass
+        # Sync forwarding rules from the fresh reload to ensure UI matches disk
+        try:
+            reloaded_new = self.find_connection_by_nickname(connection.nickname)
+            if reloaded_new:
+                connection.forwarding_rules = list(reloaded_new.forwarding_rules or [])
+                logger.info("New connection '%s' has %d rules after write", connection.nickname, len(connection.forwarding_rules))
+        except Exception:
+            pass
+        return connection
